feat: add maps from nifti and label information
This is meant to be a preparation for querying features with volumes. The previous iterations required computation of the relations to the defined maps. With ability top query with volumes, this computation is unnecessary.

In addition, this PR allows creating a map from multiple volumes and supplying parcellation specification.

TODO:
- [] a method for adding custom parcellations.
- [] a method for adding custom space.
- [] a method to extracting these custom atlas concepts as json preconfiguration